MediaRecorder: Add media recorder manual test page to testrtc.
This CL adds a manual test page for media recorder.
All the parameters of media recorder are configurable in this
test page. Also the user can choose to record the local stream or
a remote peer connection (loopback) stream. The promise version
of gUM is needed for this to work properly.